requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
astor==0.8.1
attrs==25.3.0
azure-ai-ml==1.10.0
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.13.0
azure-mgmt-core==1.5.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
backoff==2.2.1
banks==2.1.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
duckdb==1.2.2
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
Faker==19.13.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
griffe==1.7.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jpype1==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core~=0.3.58
langchain-openai==0.3.12
langchain-pinecone==0.2.5
langchain-tests==0.3.17
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.3.29
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.27
llama-cloud==0.1.18
llama-cloud-services==0.6.12
llama-index==0.12.31
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.31
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.37
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-pinecone==0.4.5
llama-parse==0.6.12
loguru==0.7.3
lxml==5.3.2
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==2.2.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
openai==1.72.0
openapi-schema-pydantic==1.2.4
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==1.5.3
pandasai==2.4.2
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pi_heif==0.22.0
pikepdf==9.7.0
pillow==10.4.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.26.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydash==5.1.2
PyJWT==2.10.1
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.25.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlglot==25.34.1
sqlglotrs==0.3.0
starlette==0.46.1
strictyaml==1.7.3
striprtf==0.0.26
sympy==1.13.1
syrupy==1.4.2
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
testfixtures==8.10.0
tika==2.7.1
tiktokapi==0.8.0
tqdm~=4.67.1
transformers==4.28.1
typing-extensions==4.9.0
uc-micro-py==1.0.1
urllib3==2.3.0
wand==0.6.10
watchdog==3.0.0
webcolors==1.12
webencodings==0.5.1
whisper==2025.6.23
xlrd==2.1.1
xmltodict==0.12.0
yarl==1.9.2
zipp==3.12.0

tiktoken~=0.9.0
fitz~=0.0.1.dev2
spacy~=3.8.5